Sample CLK *and* DATA when checking for PS/2 RTS
The CLK line is only held low for 100μs which is incredibly hard to catch. I tried solving this with an interrupt, but for some reason my system drops CLK a few times a second for no apparent reason, breaking execution far too frequently.
After probing DATA for a bit, it seems that it gets held low significantly longer and is generally more reliable than checking CLK (this issue was also expressed in Harvie/ps2dev#8).