Break backlight brightness functions down
There's no need to take an arbitrary signed integer and try process that when we already rely on a definition to change the value.
Due to an issue with the output compare module, only the output pin is disabled when disabling the backlight. This is because, if timed just right, the pin would remain toggled high when the timer is stopped and cannot (for whatever reason) be pulled low by the CPU. Disabling the output and leaving the timer running does not seem to have this problem.