Use pullup-sink method instead of pulldown
Was having trouble with the pulldown resistors being finnicky.
Instead of dealing with that, now we are reading the matrix from each column, using PORTD as a togglable sink (instead of being set high) for each row, and using ports A-C configured with pullups to measure if they have been pulled down on a keypress.
This is considerably more reliable than the previous solution and requires minimal hardware change.