Allow sending arbitrary data from host to device
Adds a new scancode that can be sent from the host: 0xCX
The upper part of the byte is always 0xC which distinguishes it from other codes, and the lower part of the byte represents the number of bytes to follow which are part of the same command. (i.e. 0xC3 will have another 3 bytes following, 0xCF will have 15, etc.)
This allows up to 16 bytes to be sent (including the initial scancode) in one go, opening up avenues for more dynamic reprogramming of the key layout, backlight, etc.