Add support for raw data over PS/2
Host can send a command (0xBB in this case) to switch the device to raw
byte mode. This will cause `keyboard_handle` to pass back a raw byte
sent by the host instead of interpreting it as a command.

True commands can still be sent by using an escape code (0xAA in this
case) which will trigger normal interpretation for the following byte.